+++ gneiss/regression/_model.py
@@ -9,10 +9,6 @@ import pandas as pd
 from skbio.stats.composition import ilr_inv
 from gneiss._model import Model
 from gneiss.util import HAVE_Q2
-if HAVE_Q2:
-    from gneiss.plugin_setup import plugin
-    import qiime2.plugin.model as model
-    from qiime2.plugin import SemanticType
 
 
 class RegressionModel(Model):
@@ -191,25 +187,3 @@ class RegressionModel(Model):
             p.name = r.model.endog_names
             pvals = pvals.append(p)
         return pvals
-
-
-# q2 cli
-if HAVE_Q2:
-    Regression_g = SemanticType('Regression_g', field_names=['type'])
-
-    class RegressionFormat_g(model.BinaryFileFormat):
-        def sniff(self):
-            try:
-                res = RegressionModel.read_pickle(open(str(self)))
-                return isinstance(res, RegressionModel)
-            except:
-                return False
-
-    RegressionDirectoryFormat_g = model.SingleFileDirectoryFormat(
-        'RegressionDirectoryFormat_g', 'regression.pickle', RegressionFormat_g)
-
-    plugin.register_semantic_types(Regression_g)
-    plugin.register_formats(
-        RegressionFormat_g,
-        RegressionDirectoryFormat_g
-    )
