+++ gneiss/regression/_regression.py
@@ -5,22 +5,57 @@
 #
 # The full license is in the file COPYING.txt, distributed with this software.
 # ----------------------------------------------------------------------------
+import pandas as pd
+import skbio
+from ._ols import OLSModel, ols
+from ._mixedlm import LMEModel, mixedlm
 from gneiss.util import HAVE_Q2
-from qiime2.plugin import SemanticType
 
-if HAVE_Q2:
-    from ._model import (Regression_g, RegressionFormat_g,
-                         RegressionDirectoryFormat_g)
-    from gneiss.plugin_setup import plugin
+from q2_composition.plugin_setup import Composition
+from q2_types.feature_table import FeatureTable
+from q2_types.tree import Phylogeny, Rooted, Unrooted
+from qiime2.plugin import Str, Metadata, SemanticType
+from gneiss.plugin_setup import plugin
+from ._type import Regression_g, Linear_g, LinearMixedEffects_g
 
-    Linear_g = SemanticType('Linear_g',
-                            variant_of=Regression_g.field['type'])
 
-    LinearMixedEffects_g = SemanticType('LinearMixedEffects_g',
-                                        variant_of=Regression_g.field['type'])
+def ols_regression(table: pd.DataFrame, tree: skbio.TreeNode,
+                   metadata: Metadata, formula: str) -> OLSModel:
+    res = ols(table=table, tree=tree, metadata=metadata._dataframe,
+              formula=formula)
+    res.fit()
+    return res
 
-    plugin.register_semantic_types(Linear_g, LinearMixedEffects_g)
-    plugin.register_semantic_type_to_format(
-        Regression_g[Linear_g | LinearMixedEffects_g],
-        artifact_format=RegressionDirectoryFormat_g
-    )
+
+plugin.methods.register_function(
+    function=ols_regression,
+    inputs={'table': FeatureTable[Composition],
+            'tree': Phylogeny[Rooted | Unrooted]},
+    parameters={'formula': Str, 'metadata': Metadata},
+    outputs=[('linear_model', Regression_g[Linear_g])],
+    name='Simplicial Ordinary Least Squares Regression',
+    description="Perform linear regression on balances."
+)
+
+
+def lme_regression(table: pd.DataFrame, tree: skbio.TreeNode,
+                   metadata: Metadata, formula: str,
+                   groups: str) -> LMEModel:
+    res = mixedlm(table=table, tree=tree, metadata=metadata._dataframe,
+                  formula=formula, groups=groups)
+    res.fit()
+    return res
+
+
+plugin.methods.register_function(
+    function=lme_regression,
+    inputs={'table': FeatureTable[Composition],
+            'tree': Phylogeny[Rooted | Unrooted]},
+    parameters={'metadata': Metadata, 'formula': Str, 'groups': Str},
+    outputs=[('linear_mixed_effects_model',
+              Regression_g[LinearMixedEffects_g])],
+    name='Simplicial Linear mixed effects regression',
+    description="Build and run linear mixed effects model on balances."
+)
+
+
