+++ gneiss/regression/tests/test_ols.py
@@ -13,8 +13,7 @@ from skbio.stats.composition import ilr_inv
 from skbio import TreeNode
 from skbio.util import get_data_path
 from gneiss.regression import ols
-from gneiss.regression._ols import ols_regression
-
+from qiime2.plugin import Metadata
 
 class TestOLS(unittest.TestCase):
     def setUp(self):
@@ -295,33 +294,6 @@ class TestOLSFunctions(TestOLS):
                   table=self.y, metadata=self.x, tree=self.t2)
         res.fit()
         self.assertAlmostEqual(res.mse, 0.79228890379010453, places=4)
-
-
-class TestOLSPlugin(TestOLS):
-
-    def test_ols_regression(self):
-        res = ols_regression(self.table, self.tree, self.metadata, 'real')
-        res_coef = res.coefficients()
-        exp_coef = pd.DataFrame(
-            {'Intercept': [0, 1.00],
-             'real': [1.0, 0]},
-            index=['Y1', 'Y2'])
-
-        pdt.assert_frame_equal(res_coef, exp_coef,
-                               check_exact=False,
-                               check_less_precise=True)
-        # Double check to make sure the fit is perfect
-        self.assertAlmostEqual(res.r2, 1)
-
-        # Double check to make sure residuals are zero
-        exp_resid = pd.DataFrame([[0., 0.],
-                                  [0., 0.],
-                                  [0., 0.],
-                                  [0., 0.],
-                                  [0., 0.]],
-                                 index=['s1', 's2', 's3', 's4', 's5'],
-                                 columns=['Y1', 'Y2'])
-        pdt.assert_frame_equal(exp_resid, res.residuals())
 
 
 if __name__ == "__main__":
